// src/generated/models/Afb_evaluationpartenairesModel.ts
export const Afb_evaluationpartenairesafb_niveauderisquecalcule = {
  0: 'Faible',
  1: 'Mod_r_',
  747010001: '_lev_',
  747010002: 'Critique'
} as const;
export type Afb_evaluationpartenairesafb_niveauderisquecalcule = keyof typeof Afb_evaluationpartenairesafb_niveauderisquecalcule;
export const Afb_evaluationpartenairesafb_statutdelevaluation = {
  0: 'Valid_e',
  1: 'Enrevue',
  747010001: 'Rejet_',
  747010002: 'Brouillon'
} as const;
export type Afb_evaluationpartenairesafb_statutdelevaluation = keyof typeof Afb_evaluationpartenairesafb_statutdelevaluation;
export const Afb_evaluationpartenairesafb_tendanceparrapportalevaluationprece = {
  0: 'Am_lioration',
  1: 'Stable',
  2: 'D_gradation',
  3: 'Premi_re_valuation'
} as const;
export type Afb_evaluationpartenairesafb_tendanceparrapportalevaluationprece = keyof typeof Afb_evaluationpartenairesafb_tendanceparrapportalevaluationprece;
export const Afb_evaluationpartenairesstatecode = {
  0: 'Actif',
  1: 'Inactif'
} as const;
export type Afb_evaluationpartenairesstatecode = keyof typeof Afb_evaluationpartenairesstatecode;
export const Afb_evaluationpartenairesstatuscode = {
  1: 'Actif',
  2: 'Inactif'
} as const;
export type Afb_evaluationpartenairesstatuscode = keyof typeof Afb_evaluationpartenairesstatuscode;

export interface Afb_evaluationpartenairesBase {
  afb_commentaireglobal?: string;
  afb_datedevalidation?: string;
  afb_datedevaluation: string;
  afb_debutdelaperiodeevaluee: string;
  afb_ecartparrapportalanoteprecedente?: number;
  "afb_evaluateur@odata.bind": string;
  afb_evaluationpartenaireid: string;
  "afb_evaluationprecedente@odata.bind"?: string;
  afb_findelaperiodeevaluee: string;
  "afb_grilleutilisee@odata.bind": string;
  afb_niveauderisquecalcule: Afb_evaluationpartenairesafb_niveauderisquecalcule;
  afb_noteglobale: number;
  afb_notemaximalepossible: number;
  afb_plandactions?: string;
  afb_referencedevaluation?: string;
  afb_statutdelevaluation: Afb_evaluationpartenairesafb_statutdelevaluation;
  afb_tendanceparrapportalevaluationprece?: Afb_evaluationpartenairesafb_tendanceparrapportalevaluationprece;
  "afb_tiersevalue@odata.bind": string;
  "afb_validateur@odata.bind"?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Afb_evaluationpartenairesstatecode;
  statuscode?: Afb_evaluationpartenairesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Afb_evaluationpartenaires extends Afb_evaluationpartenairesBase {
  afb_evaluateurname?: string;
  afb_evaluationprecedentename?: string;
  afb_grilleutiliseename?: string;
  afb_niveauderisquecalculename?: string;
  afb_statutdelevaluationname?: string;
  afb_tendanceparrapportalevaluationprecename?: string;
  afb_tiersevaluename?: string;
  afb_validateurname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  afb_evaluateur?: object;
  _afb_evaluateur_value?: string;
  afb_evaluationprecedente?: object;
  _afb_evaluationprecedente_value?: string;
  afb_grilleutilisee?: object;
  _afb_grilleutilisee_value?: string;
  afb_tiersevalue?: object;
  _afb_tiersevalue_value?: string;
  afb_validateur?: object;
  _afb_validateur_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}